🛠️ Refactor: putting the application in the components
