+++ frontend/src/components/TransactionManagement/LoadingPage.tsx
@@ -1,0 +1,9 @@
+export const TransactionManagementLoadingPage = () => {
+    return <div className="flex justify-center items-center p-4 min-h-screen">
+        <div className="flex flex-col items-center space-y-4 p-8 rounded-lg w-full sm:w-96 max-w-md">
+            <div className="border-gray-900 mb-4 border-b-2 rounded-full w-32 h-32 animate-spin"></div>
+            <div className="font-semibold text-gray-700 text-xl">Loading...</div>
+        </div>
+    </div>
+
+}
